Modified ShowData method in display package - argument changed from float to double.
Created input package - its purpose is to catch info from the user.

Main class - added main loop (which is still not complete), added methods asking and retrieving data from the user. Added methods that push calculated data to the user.

+++ src/lab1proj/display/ConsoleDisplay.java
@@ -13,7 +13,7 @@ public class ConsoleDisplay implements IDisplayModule
     }
 
     @Override
-    public void ShowData(float number) {
+    public void ShowData(double number) {
         System.out.print(number);
     }
 }
